FEAT : created apis for login, show courses and enroll course

+++ ESD_MINI/src/main/java/com/maitri/esd_mini/repo/StudentRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 import com.maitri.esd_mini.entity.Students;
-public interface StudentRepo extends JpaRepository<Students, Long> {
+public interface StudentRepo extends JpaRepository<Students, Integer> {
     Optional<Students> findByEmail(String email);
 }
